[pkg/util] Cleaned up `SlidingWindow` overlapping lock usage
Since the two sets of locks in `slidingWindow` were added at different
times, their usage wasn't as well coordinated as it needed to be. By
moving some code around and ensuring that we lock minimal sections of
the code, we now should have no overlap in the two lock footprints.