[flare] Include logs with the existing subdirectory tree
If the user creates subdirectories in the agent's log directory,
and for example moves agent logs to these subdirectories, this
change allows the directory tree to be preserved in the flare.

Previously, in this scenario, log files in these subdirectories
could overwrite the log files in the root log directory
when they have the same basename.