tagger: convert pod ids between tagger and kubeUtil
The kubernetes-metadata-collector used to call kubeUtil during a fetch
without converting its own IDs (kubernetes_pod_uid://) into the ones
that the kubeUtil understands (kubernetes_pod://), causing a failure for
every fetch of a pod. But not a container! They way GetPodForEntityID
works is to check the prefix to see if it matches that of a pod, and
assume it's a container otherwise. Since we were passing what's
essentially a malformed ID, it always assumed containers for either pods
or containers.