[flare] Include logs with their existing subdirectory tree (#9042)
If the user creates subdirectories in the agent's log directory,
and for example moves agent logs to these subdirectories, this
change allows the directory tree to be preserved in the flare.

Previously, in this scenario, log files in these subdirectories
could overwrite the log files in the root log directory
that had the same basename.

Also, add the file permission information on the log directory
itself (e.g. `/var/log/datadog`), which were missing.