[tests] Exclude flappy `TestUtilizationTrackerAccuracy` from Windows tests
Timing-based tests on Appveyor seem to be generally flaky and while this
test in most cases passes without issues, the false negative case is
still high enough of a probability to be an annoyance. Since this test
uses no OS-specific functinality and passes on Linux with much higher
confidence, we still have assurance that this logic works as expected.