Add support for remote config
This PR adds support for requesting configurations from the backend.

Client can subscribe to configuration change to a service. This service will
collect all the configurations to fetch and periodically issue a request to
the backend. The response is then verified according to the TUF specification.
The configurations are then dispatched to the subscribers. These subscribers
can be local subsystems or remote gRPC clients.

This feature is disabled by default.