[tooling] Add `dep_why.py` tool to help identify dependency links
This tool automates a common use case of the dep_tree_resolver tool
output by scanning the generated dependency tree and printing out any
tree branches that lead to the specified module name. This is useful in
identifying the reasons why a particular module is being used and
through which hierarchy of dependencies.