Add support for requesting remote config (#9253)
* Add support for remote config

This PR adds support for requesting configurations from the backend.

Client can subscribe to configuration change to a service. This service will
collect all the configurations to fetch and periodically issue a request to
the backend. The response is then verified according to the TUF specification.
The configurations are then dispatched to the subscribers. These subscribers
can be local subsystems or remote gRPC clients.

This feature is disabled by default.

* Update protobuf generated files

* Remove unused SIGHUP subscriber

* Fix android build

* Allow setting refresh interval

Also changed the default to 1 minute and check against a minimum value

* Make an empty hash list an error

* Allow closing grpc subscribers

* Report config delegated version to subscribers

* Move configuration to a remote_config section

* Disconnect subscriber when we fail to send it a message

* Use configuration to get local gRPC endpoint

* Improve GetMeta documentation

* Log errors in pruning bucket only once

* Add config product and director root version to GetConfigsRequest

* Do not advertise command

* Remove useless map and array replaces

* Use run_path configuration know for database path

* Sleep a bit to avoid log flood with incorrect config