Always use atomics to access an atomic value
The existing implementation used a "normal" read to read a value,
increment it, and then write it atomically.  This mixed "normal" and
atomic reads of the same value, and also did not accomplish the intended
atomic increment.

In this case, because this operation is meant only to distribute the
load over a number of pipelines, the result of any problems here would
be, at worst, uneven distribute of pipelines.  Still, let's not keep
incorrect implementations of low-level operations around in the
codebase.

This changes from int32 to uint32 since it's used as an array index
which must be nonnegative.