Factor tagset hashing to pkg/tagset
This is part of ongoing efforts to centralize tagset-related
functionality in pkg/tagset, as a generalized utility for all agent
functionality.

This commit replicates the existing use of a struct field instead of a
register for the hash accumulator.  Some disassembly suggests this is
inefficient, and does not avoid the stack-length check on the Generate
function.

This also replicates the use of an initial, arbitrary hash value.  We
could save one XOR operation by just beginning with murmur3(name).