Fixing deadlock in the metadata scheduler
When stopping the metadata scheduler we lock its main mutex and then
stop the job queues. The issue is that the job queues might call
'scheduler.IsCheckScheduled' which tries lock the main mutex again.

Deadlock order when stopping the scheduler:
1. We lock the scheduler mutex.
2. The jobQueue call 'IsCheckScheduled' trying to lock the mutex and blocking for ever.
3. The scheduler hang forever trying to stop the job queues who are
   blocked on the `IsCheckScheduled' mutex.

Adding a secondary RWLock to around 'checkToQueue' solves the issue.