Probe the flare endpoint with HEAD
The standard URL for flares is handled with a 307 redirect to another
URL.  Large POST bodies and 307 redirects don't play nicely together:
servers are free to reply and hang up the connection before reading the
request body, and net/http doesn't always recognize this situation
properly, resulting in spurious errors.

At any rate, it's a waste of resources to send a multi-MB zip file to an
edge endpoint that is not interested in that data.

So, this PR adjusts the flare-sending functionality to use HEAD requests
to evaluate the redirect chain, and only POST to the final URL.  This
generates one additional HEAD request (to determine that the redirect
chain has ended), but overall represents a savings in traffic.