Add expvars to measure logs agent sender utilization (#9412)
This change exposes a new group of expvars for batch_strategy. The metrics can be used to measure the logs pipeline sender utilization. These are most useful when paired with the go expvar check (default config is also updated).

You may compute the utilization in app with: inUseMs / (idleMs + inUseMs) and break it down per sender to understand how saturated the senders are and if more concurrent senders need to be enabled.