Remove unused embedded interface (#9495)
The syntax

```go
type struct NewLineMatcher {
    EndLineMatcher
}
```

where EndLineMatcher is an interface type, simply embeds a
EndLineMatcher into Foo.  In this case, that embedded EndLineMatcher is
never initialized, so calling any of its methods would fail with a
segfault.  But all of its methods are redefined for NewLineMatcher
anyway.  So the interface is just a wasted 16 bytes of 0's.