[tagger] Stop stream when idle clients disconnect (#9452)
This prevents a goroutine leak in environments that don't produce any
tags over long periods of time. Waiting on eventCh would not check on
the state of the stream until an event came, so if the client
disconnected due to inactivity (which is currently every 10 minutes), a
new goroutine to handle the stream would be established, but the old one
would remain open indefinitely.

Now, we select on both eventCh, and the context of the underlying
stream. Once a client disconnects, that context is cancelled, and our
goroutine can exit cleanly.