[ecs] Remove metadata.V3()
This method used to call into dockerutil to get a container's
environment variables to be able to find out which API endpoint to use.
Although the need for such a process still exists, it has been migrated
to use workloadmeta's store, which avoids calling into the docker
daemon, and all callers had been previously migrated as well.