[PodWatcher] Generate events for ready->unready->ready pods
Previously, once a pod was observed as ready, going unready and back to
ready would not re-schedule checks, making previously running ones go
away, with an "unable to resolve, service not ready" error and never
come back as the pod became ready again.

This commit also addresses the fact that a pod could now generate many
services for the same pod. Now we also check that pod services aren't
duplicated, just like we do with containers.