[workloadmeta] Separate pulling and event processing
This applies the same strategy as #8830, for a related (albeit much
rarer) reason. The store could still deadlock if the initial
`startCandidates` held a lock on collectorMut for long enough to a
`pullTicker` to trigger, and caused enough events to be generated to
fill up `eventCh`. In that case, `startCandidates` will be blocked on
any sends to `eventCh`, and the store goroutine will be stuck `pull`
waiting for `startCandidates` to release its lock.

As a bonus, this also fixes a data race on `pullCtx` and `pullCancel`
being written to (in the <-retryTicker case) while they're being read
(in the main goroutine).