[kubernetes] [clustername] handle some validation edgecases (#9279)
In the past we were not able to send some hostAlias for autodetected kubectl hostNames. As they are built on the clusterName. If the clusterName is not valid RFC1123 then the agent did not send it as a hostAlias.
EKS, AKS cluster names are almost RFC1123 compliant. Except for those cases:

ending with - or _
having _
This PR makes sure to catch most of those cases by replacing _ with -. With this our backend is now able to send a hostAlias for kubernetes i.e.
clusterName: my_cluster
host: i-321312312

The resulting hostAlias will be
my-cluster-i-321312312

Additionally this information will be shown in the datadog-agent status output