[tagger] Prevent workloadmeta fetches from overwriting good tags
This fixes a race condition where tagger.Tag finds no tags for an
entity and triggers a Fetch, while at the same time the workloadmeta
collector is processing a batch of events that would've satisfied the
query. Whenever the Fetch wins the race, it overwrites a valid set of
tags with empty tags.