Add saturation metrics for network and memory
We currently gather utilization and error metrics of network (bytes received and sent, number of packets) but not the saturation metrics (number of dropped packets). Similarly for memory, we gather the utilization metrics (free, used) but not the saturation (page in/out, swap in/out).

Having this additional data is needed to apply the USE Method [1] for system performance analysis.

[1] https://www.brendangregg.com/usemethod.html