[github] Make default PR template section descriptions be invisible by default
Currently, when a PR is opened and the default template text used without
modifications, the PR may look like it has relevant information in those
sections when in actuality it's just the template content.

This change ensures that it's obvious to the reviewer(s) when the PR content
has not been filled out by the author(s).