Fix deadlock in WeightedQueue
The deadlock was due to use of an unbuffered channel as a
condition variable.  It could be reproduced by adding a `time.Sleep`
call before the check of `q.dataAvailable` in `Poll`.  The timeline is:

 * [goroutine 1] calls Poll()
 * [goroutine 1] locks q
 * [goroutine 1] sees q.queue.Len() == 0
 * [goroutine 1] unlocks q
 * [goroutine 2] calls Add
 * [goroutine 2] locks q
 * [goroutine 2] pushes a value into the queue
 * [goroutine 2] tries ->q.dataAvailable, but finds no pair and continues
 * [goroutine 1] executes <-q.dataAvailable and hangs

An easier fix might be to use a buffered channel of size N, where N is
the maximum number of concurrent Poll operations.  Current usage of
WeightedQueue suggests N=1, but that is a difficult API constraint to
place on users of this type.

So the fix applied here is to use a condition variable which is
signalled both when there is data available in the queue and when the
queue is stopped.  This simplifies the Poll operation, but requires a
simple goroutine to "translate" `<-exit` into `wq.Stop()` calls.