Enable dual shipping of logs to a secondary endpoint (#9670)
This change enables dual shipping of logs. Add an additional sender with is_reliable: true and the agent will start sending logs to that org/endpoint as well. This additional endpoint wills still be sent to if the main endpoint goes down. Messages will be retried on retryable errors, back-pressure to the input will be preserved under slow/irregular conditions, and the auditor will be updated to ensure no duplicate logs are sent, and no logs are missed.