[workloadmeta/ecs] Avoid calls to metav3 for known tasks
This is a forward-port of #9902, to prevent the agent from doing too
many calls to the ECS V3 API and getting throttled, and therefore not
being able to track tags for tasks. This PR follows the same strategy as
the old tagger collector, by considering that once a task starts
running, it won't change, and therefore there is no need to continuously
generate Set events for it.