[AD] Track children services before skipping duplicates
This fixes an issue where a container inside a pod would have its
services removed if any other service related to the pod was added,
changed or removed. This was very apparent whenever a container had a
check, another container in the pod was restarted, and all checks would
become unscheduled.

The issue happened because a change to a pod would cause all containers
to be re-inspected, and if they hadn't changed, they'd just get
unscheduled becase we skipped duplicated services before tracking the
parent-child relationship. We solved it by moving the tracking to before
the point where we do the skipping.