fix: Add missing node tag to kubernetes_state.pod.count metric
the `kubernetes_state.pod.count` metric is computed from the `kube_pod_info`
that contains the label `node`. to be able to map the `kube_pod_info`
to the host, the `pod.count` aggregator need to use the `node` label.

Add missing `tag` reference in the `kubernetes_state.md` documentation.