Add container_mode_docker_origin config option to journald log collector
The existing container_mode option causes journald logs to have an
origin source and service set to the Docker image short name. However,
this takes precedence over service tags derived from the container env
vars and/or labels, which makes it impossible to override the service
though the usual mechanisms.

This change adds a "container_mode_docker_origin" option which makes the
journald collection act like the direct docker log collector when in
container mode by setting the source to "docker" and not explicitly
setting a service, which allows it to be set via tags.

Although the name might be a little confusing, removing or changing the
existing container_mode option could be considered a breaking change.