[tools] Add copyright linting tool for Golang files
This addition can be used via `inv lint-copyright` which traverses the repo
tree and fails if it finds any Golang files with an unexpected copyright
header.

For now the tool is intended to be invoked manually but once all the files
get fixed, we can add it to our internal CI/CD pipeline as an automated
check.