[tools] Add basic fixing capabilities to the copyright linter tool
Previously, the `inv lint-copyrights` tooling was only able to detect
problems with the copyright headers. This change adds functionality to
fix the vast majority of these issues (specifically when the header is
completely missing) when run with `--fix` flag. A `--dry-run` flag was
also added as a way to safely test run this command before making
permanent changes to files.