Backport of: (#10203)
Incorporates new version of agent-payload, which fixes the following
bug:

The offset of the middle entry of the dns database is being encoded,
but not used (was originally left for future optimization).
However, it has a problem of the result changing its own position.
This works, except in a very specific boundary case.

if offsetOfMiddle happens to be 127 (or any other subsequent size
that causes the size of a varint to go up), we have an off-by-one bug.
The offset is 127, so we compute the size (which is 1), and then
increment
the buffer size to match. However, since the offsetOfMiddle is now 128,
the
size of the varint is now 2, and the whole buffer's whacked. Only when
the
middle happens to be on the boundary of when the varint size changes.