[tagger] Be more resilient when notifying remote tagger
This prevents the tagger from being slowed down by remote tagger
subscribers that are no longer responding timely, by cancelling their
subscription entirely if their channel fills up. If the subscriber is
still healthy, it'll retry and get a fresh subscription, but more likely is that
the subscriber is not healthy, so at least it won't slow down the
tagger anymore. Reminder that it's important that the tagger processes
events as fast as possible, as otherwise other components depending on
the workloadmeta won't get their timely updates, as they're all
downstream.