[tagger] Remove tagstore.InfoIn
The tagger no longer needs to run tag processing in parallel, since all
tags come from a single source, and those are processed sequentially. To
reduce complexity, and fix the one remaining known race condition in the
tagger, where closing the `workloadmeta.EventBundle` channel did not in
fact guarantee that the tags were in the store.