[workloadmeta] Refactor source filtering
Previously, each collector pretty much owned its own `Source` (as in,
the `kubelet` collector used a `kubelet` source, the `docker` collector
used a `docker` source, and so on). This implies that components that
filter by container runtimes would need to update their filters for
every new container runtime the workloadmeta supports, which is less
than ideal.

To solve that, the sources have been reworked. Now we have the following
sources:

* SourceRuntime: entities detected by the container runtime running on
  the node, collecting lower level information about containers.
  `docker`, `containerd`, `podman` and `ecs_fargate` use this source.
* SourceNodeOrchestrator: entities detected by the node agent from an
  orchestrator. `kubelet` and `ecs` use this.
* SourceClusterOrchestrator: entities detected by calling the central
  component of an orchestrator, or the Datadog Cluster Agent.
  `kube_metadata` and `cloudfoundry` use this.

This commit also removes the ability to filter by more than one source,
as that's no longer necessary, simplifying the code a lot.