[docker] Fix image tags for stopped containers
This fixes an issue where a stopped container that has a SHA256 as its
image name would not be counted in `docker.containers.stopped` and
`docker.containers.stopped.total` metrics. It also lowers the log level
for the `Cannot split the image name` error, since it is not actionable
and would only be useful for debugging.