[workloadmeta] Fix race condition on Subscribe
This fixes a race condition that would make a new subscriber to lose
events that happened between taking a snapshot of the current state of
the store, and adding the subscriber to the list of subscribers of the
store. To fix this race condition, we now lock the store until the
subscriber is added to the list, which blocks any new events from being
processed by the store.