[workloadmeta] Gracefully stop on agent shutdown
This passes `common.MainCtx` (where available) to
`common.LoadComponents`, which in turn passes that context to the
workloadmeta, so it can be notified that an agent shutdown has been
requested, and we can do it cleanly. Besides being just common good
practice, we're also introducing this to help debug a panic in the
tagger that's suspected to happen on shutdown, and being able to
actually request one is useful.

Note that this has only been done in the main agent for now.
Workloadmeta is not enabled by default in other agents, so there's lower
priority in doing that.