Add dogstatsd arm64 binary, .deb and docker builds
Note: .deb and docker builds are promoted, binary builds are not.

This also renames the binary build jobs from *-deb* to *-binary*, even if
they run on Debian to avoid confusion with the .deb package build jobs.

The deploy jobs for dogstatsd and iot-agent binaries now explicitly depend
on the respective binary build jobs, so they only fetch those artifacts.