Fix handling of series with too many points.
This adds a new test case provided by @ogaca-dd, and fixes the bug
causing it to fail.  The bug was using `pointsThisPayload` where I should
have used `len(serie.Points)`.

Having fixed that, the code was still flushing an empty payload when it
stopped, so the seriesThisPayload variable tracks that and skips
flushing when it is zero.

In the process of making that change, a third bug became apparent: when
the Compressor refuses a series, the existing code was still "counting"
the points in `pointsThisPayload`.  This would have had little effect,
but is remedied by only incrementing `pointsThisPayload` and
`seriesThisPayload` when the compressor accepts a series.