[inv deps] Fetch dependencies for all submodules
Some nested modules use specify versions of dependencies that differ from
one module to another. Althoughh at build time only one version will be
used, when running commands on a specific module (which we do when linting,
for example) the version specified locally on the module will be used.
This change ensures those dependencies are pre-fetched.