Convert pkg/logs/parser.Parser#Parse to return a struct, rather than a 5-tuple
This is a formaulaic change, although the added documentation is
important.

The performance implications of this change are minimal, but nonzero --
Both returned structs and tuples are stored on the caller's stack frame,
but Go moves a struct return value after return whereas it just uses a
tuple return in place.  This seems like something the compiler may
eventually fix!

At any rate, this is not an especially performance-critical part of the
code, and is surrounded by much less performant things.