Docs / naming updates for pkg/autodiscovery (#10520)
* update README with current structure

* update Autoconf doc comment, removing old info

* Remove outdated comment

This comment was added to document a code change in
b93e1f4e30a436a50c6314e42605864d1bf054e2.  The code has since been
reverted.

* Describe what 'loaded configs' means

* make pkg/autodiscovery.TemplateCache private

* doc ConfigProvider methods

* integration.Config docs + drop unused method

The Type() method is unused, and did not distinguish template,
non-template, and service configs.

* make ProviderCache private, doc it, and use clearer field names

* include cluster/endpoint checks and kube_*

* mention template variables

* fix typo

* Rename AD 'entity' to 'service ID'

This field is, in fact, a unique identifier for the AD service and is
not to be confused with entities as defined by either the tagger or the
workloadmeta service.  Renaming the helps avoid that confusion.

* don't use containers.BuildEntityName for things that aren't containers

* clarify Config.Entity vs. Config.TaggerEntity

* Rename integration.Config.Entity to ServerID

This better reflects its meaning, as it does not correspond to either a
WLM entity or a Tagger Entity.

* do not treat serviceIDs as entity IDs

* fix missed rename

* fix incorrect doc comments in tests

* fix README typos