[AD] Prevent workloadmeta listeners from hanging at shutdown
When the agent shuts down, one of the first things it does is to cancel
the `MainCtx`, that since recently is propagated to the workloadmeta
store, so that it can shutdown as well. As workloadmeta shuts down, it
cancels all of its subscriptions, and in AutoDiscovery that means that
listeners shut down with it.

While all of this is happening, the agent also tells AD to stop itself,
and part of that is to tell each listener to stop with it. Since the
workloadmeta listeners may have already shut down, a send on the
unbuffered `l.stop` hangs, as the receive loop has stopped. This makes
the entire agent shutdown process to hang as well.

Since there's exactly one call to `Stop`, it's safe to just close the
channel. As it's actually *not* expected to be called more than once,
the possible panic that a second call can cause is actually desired so
the bug is apparent.