[Remote Config] Add exponential backoff to HTTP client (#10760)
Adds exponential backoff to the remote config HTTP client to
prevent agents from overwhelming the remote config edge during service
delivery issues.

To do this, we leverage pkg/util/backoff and define a `backoff.Policy`
for the remote config client. We keep track of the current error level
on successful and failed updates, and calculate a refresh interval for
each request to the remote config backend based on the state of the
system.

The backoff policy has jitter built in for random backoff times. The
default values we use for the policy start with a backoff duration that
will be within the range [30s, 60s], with the min and max doubling with
every consecutive error. When we see a success, we drop back two
states. (equivalent to dividing the min and the max by 4) The maximum
backoff added is is configurable and is constrained to be between 2m
and 5m. No other values related to the backoff policy are configurable
by the user at this time.

Resolves: [RCM-58]