[tagger] Prevent panic on double closed channel
When a remote tagger is unable to receive updates from the core agent,
its channel fills up and gets closed automatically by the upstream. When
that happens, the subscription in the gRPC server also unsubscribes,
resulting in a double close. To prevent that, we check that we still
have the channel in the list of subscriptions instead of blindly closing
the passed in `ch`.