Store metric and tagger tags separately
This patch is based on observation that tagger tags usually have lower
cardinality than metric tags. If stored separately, tagger tags can be
shared between higher number of contexts, reducing number of tags in
the per-context slice (which can also be shared better, if the same
tagset happens to come from different origins).

Use separate accumulators for tags that came with the metric sample
and for tags from the tagger (origin detection). Use Hash2 to
deduplicate them together and use hashes to store them separately in
the store. Leverage CompositeTags to pass the two tags slices
separately all the way to the serializer without having to allocate
unified slice most of the time.