[workloadmeta/ecs] Re-parse tasks on every pull
This fixes a regression introduced by #9928. While we wanted to prevent
the ECS collector from hitting the metav3 endpoint on every task update
to avoid hitting throttling limits, we also prevented updates to other
properties of the tasks from being processed. We expected such
properties (specifically, the list of containers) to be immutable, but
we since learned that's not the case: containers get added to the list
as they start, not atomically when the task gets added in the API.

No, we re-parse all tasks on every pull, but as noted in the comments,
we don't re-fetch AWS tags to not hit throttling limits again. This is
an acceptable trade-off as this matches the behavior in the Agent <7.32.