[workloadmeta/containerd] Use data already present in Info
When running the containerd collector, we noticed that we get errors
like `could not fetch info for container: failed to get image $IMAGE for
container: image "$IMAGE": not found`. This is because we try to fetch
all of the image information from the containerd daemon, instead of
using just the image name that we already have in memory.