[pkg/forwarder] Improve error messages for http errors in API key validity check
When using the `dd_url` option and going through a misconfigured proxy, then
hitting a redirect endpoint may make the agent connect directly to the redirect
target. When this occurs, debugging may be difficult as the target
redirect domain is shown in the error messages instead of the original
`dd_url` domain. By adding this info to the error message, exact
pinpointing of the problem can be made quicker.