Minor updates to logs-agent tailers and launchers (#10970)
* remove unused field from docker tailer

t.shouldStop is only ever written, so is unused.  It used to be used to
avoid re-checking container tags, but that functionality is no longer in
the tailer.

* Update docker tailer comments, names

Especially around how the tailer stops.  This is a non-functional
change, although the renames make the diff look rather large.

* Document docker.Tailer#ContainerID

The docker launcher works in terms of container IDs, so it needs access
to this value in order to do its work.

* file tailer: ShouldStop -> IsFinished
* file tailer: inline some methods
* file tailer: simplify didFileRotate
* remove outdated functional diagram