Fix metric confusion when the metric name is contained in the tags
Previously, when the metric name occurred in the tags list, the two
hashes would cancel out when we combine them with XOR, causing the
same context key to be returned for different metrics.

Instead of using XOR, chain hashes by feeding previous output as a
seed value for the next.