pkg/obfuscate: fix parsing negative numbers being mistakenly interpreted as octal
Any number matching -\d0\d*[89]\d* format was misrecognized as an octal number due to the parser ignoring the first digit.

Additionally Any number -.\d+ was not interpreted as a single number, but as distinct tokens '-' and a number. Added an additional case to handle this scenario to ensure proper normalization.