Handle windows log file rotation using DidRotate
Previously, there was a bit of a "hack" where the tailer, on Windows,
would quietly reset its offsets to 0 when it detected a file truncation.
This avoided the entire file-rotation process that involves the launcher
creating a new tailer.

With this change, the same logic is used, but this time to detect a
truncation and report it to the launcher via DidRotate, just like on
*nix.  The read routine just returns no data when the file has shrunk,
again matching the results on *nix.