Windows file tailer: do not continue polling after rotation
Because the agent cannot hold an open file handle on Windows, once a
file is rotated there is nothing the agent can do to recover additional
data.  At first, the file size will be less than the last read offset,
but this might not be the case for the (default 60-second) time the old
tailer continues polling.  So, this detects the condition and
short-circuits to returning no additional data after a rotation takes
place.