Make the DEB package datadog-signing-keys a hard dependency
We added this package as a "soft" (recommended) dependency, but on some
setups this might not be enough to make sure the package is installed. For
example: if the datadog-agent package is not installed from our repos, the
missing datadog-signing-keys package won't make the installation fail.

Since 7.35 will be the last agent package we release before the key
rotation, we should include this change in 7.35 to make sure more people
install the datadog-signing-keys package.