Handle windows log file rotation using DidRotate (#11135)
* Handle windows log file rotation using DidRotate

Previously, there was a bit of a "hack" where the tailer, on Windows,
would quietly reset its offsets to 0 when it detected a file truncation.
This avoided the entire file-rotation process that involves the launcher
creating a new tailer.

With this change, the same logic is used, but this time to detect a
truncation and report it to the launcher via DidRotate, just like on
*nix.  The read routine just returns no data when the file has shrunk,
again matching the results on *nix.

* Windows file tailer: do not continue polling after rotation

Because the agent cannot hold an open file handle on Windows, once a
file is rotated there is nothing the agent can do to recover additional
data.  At first, the file size will be less than the last read offset,
but this might not be the case for the (default 60-second) time the old
tailer continues polling.  So, this detects the condition and
short-circuits to returning no additional data after a rotation takes
place.

Co-authored-by: Brian Floersch <brian.floersch@datadoghq.com>