[workloadmeta] Fix race condition when reading from the store
When notifying subscribers after handling an event, we read from the
store without having a lock on `s.storeMut`. This may result in
inconsistent data being fed to subscribers, but more importantly, it may
generate `nil` entities that can cause panics downstream (such as in
issue #10716, which we expect to be fixed in #10804, but the actual root
cause is the unsafe read) when handling deletes in succession from
different collectors.