Replace the container launcher with feature detection.
Rather than waiting for one of Docker or Kubelet to start up, this just
consults the enabled features on the agent.  It keeps the existing
preference pattern when both are enabled, based on the config
`logs_config.k8s_container_use_file`

Note that the Docker launcher is really a "container launcher", handling
all the supported container runtimes.