[tagger] Add task_* and static tags to OrchestratorScopeEntityID
This fixes missing static (aka DD_TAGS) and ECS-related tags in ECS
Fargate environments when using dogstatsd metrics and origin detection,
by fixing a regression introduced by #9557 in 7.33 where the removal of
the `static` tagger collector stopped adding these tags to
OrchestratorScopeEntityID (but not any of the containers).

This also makes OTLP honor the passed in cardinality, as it was
previously hardcoded to `orchestrator`.