[tagger] Collect static global tags in every environment
This change collects static tags (the ones returned by
GetStaticTagsSlice, ie. ones that don't change over the lifetime of the
agent) into the recently renamed GlobalEntityID as low cardinality tags,
untying them from the specific entities they used to live with
(containers, pods, tasks).

This is a no-op in most environments, except that in EKS Fargate
entities that do not resolve to a container ID (for instance, checks
configured by a file instead of AD) will also be tagged with DD_TAGS and
EKS Fargate node and cluster by the tagger itself, instead of by the
ad-hoc changes introduced by #10952 and #10983.