[clusteragent/admission/webhook] Add selectors needed for AKS
This fixes a problem that happens when enabling the DCA admission
controller on Azure AKS.

AKS automatically updates our webhooks adding a selector that excludes
the "control-plane" namespace. This creates conflicts when we reconcile
the webhooks in the controller and spams the logs with messages like:

`Couldn't reconcile Webhook datadog-webhook: Operation cannot be
fulfilled on mutatingwebhookconfigurations.admissionregistration.k8s.io
"datadog-webhook": the object has been modified; please apply your
changes to the latest version and try again`.

References:
- https://docs.microsoft.com/en-us/azure/aks/faq#can-i-use-admission-controller-webhooks-on-aks
- https://github.com/Azure/AKS/issues/1771