Stop using ipconfig on Windows in favor of golang net package
Using `ipconfig` on windows does not work when the locale is not english.

For example on ja-JP system:
```
PS C:\Users\Administrator> ipconfig

Windows IP 構成

イーサネット アダプター イーサネット:

   接続固有の DNS サフィックス . . . . .: eu-west-3.compute.internal
   リンクローカル IPv6 アドレス. . . . .: abc0::abcb:abc3:abc7:abc0%4
   IPv4 アドレス . . . . . . . . . . . .: 1.2.3.4
   サブネット マスク . . . . . . . . . .: 255.255.240.0
   デフォルト ゲートウェイ . . . . . . .: 1.2.3.4
```

The orignal commit doesn't explained why ipconfig was used in the first
place. Best guess is that the original version was introduced because in
Golang 1.2 `net.ParseCIDR` returns `nil` on Windows.

Nowaday there is no reason not to use the generic implementation from
Golang. This also aligned the format between Windows and Linux.

There is two differences in the output between ipconfig and the net package:
- ipv6 format no longer has the `%` part with the scope ID.  old
  `aaaa::bbbb:cccc:dddd:eeee%11` vs new `aaaa::bbbb:cccc:dddd:eeee`
- `macaddress` is now capitalized and use dash. Old `00:0c:29:0a:3d:a3`
  vs new `00-0C-29-0A-3D-A3`

Code was tested from 'Windows 2008 R2' to 'Server 2012'