[ksm] Make time based metrics comparable.
Prior to this patch the ksm collector would call
`time.Now().Unix.()-metric.Val` at every point a `_time` metric is
computed, this leads to the values collected in the same loop not being
comparable since their values skews based on the value of `time.Now()`
at the time metric was collected.

For an example of the impact of the above, to compute how long it takes a
pod to start after being created we create a function:

```
avg:kubernetes_state.pod.age{pod_name:filebeat-lf6hz} by {pod_name}
- avg:kubernetes_state.pod.uptime{pod_name:filebeat-lf6hz} by {pod_name}
```

When this is plotted on a dashboard we end up with values fluctuating
between negative and positive.

To ensure the values are comparable the `currentTime` is snapshotted at
the start of the Run loop and the value is reused in all places that had
previously called `time.Now()`.