Revert "Determine the container runtime from the environment. (#11101)"
This mixes up the older, hacky way of supporting Podman with the newer.  The old way was to run the agent in a podman container and set DOCKER_HOST to the podman socket.  In this arragement, the Agent has no idea Podman is running, which is fine _except_ that the log format is different -- hence the need for `logs_config.use_podman_logs`. But PR #11101 expects the agent to "know" it's running in Podman to set the appropriate feature.

The new (as-yet undocumented) way is to treat podman as a different runtime from docker, meaning that integration.Config's have a ServiceID of the form `podman://<sha>`.  In this circumstance, the features are set correctly, but the logs-agent will create a source with `Config.Type = "podman"` which will not result in any logging.

This reverts commit 94cf95cec8b3bfb21f64dd1cf99f72720c8d17de.