Create a new LogSource when changing serverless tags (#11270)
* Create a new LogSource when changing serverless tags

The "channel" LogSource is only used by the serverless agent, but could
be useful for other purposes, so this also abstracts the
serverless-specific bits (name, source) out.

More importantly, when the tags are updated, this now drops the old
source and creates a new source.  This avoids race conditions with
modifying a slice pointer while it might be in use by another gorotuine.

Since the tailers for both sources are reading from the same channel, no
messages will be lost -- a given message will either be handled by the
old tailer or, if it stops before the message is handled, will wait for
the new tailer to start.

* Protect channel tailer tags with mutex

The channel tailer closes its input channel, and the channel scheduler
does not support removing sources, so this implements the async update
of the tags applied to the channel with a new field (ChannelTags)
protected by a new mutex (ChannelTagsMutex).  The new field avoids the
requirement to guard _every_ access to LogConfig.Tags with a mutex.

* Do not multiply append tags to serverless messages.

* deduplicate ddtags in serverless snapshots

* Fix Java integration tests

* update comment

Co-authored-by: Nick Hinsch <nick.hinsch@datadoghq.com>