[demultiplexer] removes `Aggregator()` func from `Demultiplexer` interface.
The `Aggregator()` getter has been kept for a while in order to simplify the
migration from the `BufferedAggregator` to a `Demultiplexer` in multiple parts
of the code. We can now remove it more easily that in the past and this is what
this commit is doing.

I moved the aggregated data print functions out of the cmd/check part since it
is about printing data embedded in the BufferedAggregator.
A next commit will refactor this part in different files and a better structure.