[tagger] Add static tags to container/pod entities
These were removed in #11293, and then lived just in GlobalEntityID,
with the assumption that all data coming out of the agent would call
`tagger.EnrichTags`, where we add GlobalEntityID's tags to another set
of tags. That assumption is wrong, as it turns out that process and
container collection (and possibly other components) *do not* call that
func, and therefore would not have the static tags added.