[workloadmeta] Send last known state on EventTypeUnset
This relieves collectors from having to cache information right before a
container is removed, to be sent later with an EventTypeUnset. One
example of this is the containerd collector caching a task's exit code
when the task gets deleted, right before the container itself is
deleted, and the exit code information has already been lost (see
PR #10011).