Factor out config management from AutoConfig
This factors out the config management -- handling added/deleted configs
and services -- from AutoConfig into a dedicated type.  The type is an
interface implementation to support its later wholesale replacement
based on a feature flag.

This is a complex diff to read, but at its core it adds new types and
then "lifts" the body of four methods out of AutoConfig and puts them
into the simpleConfigManager type.

 * Change processRemovedConfigs to distinguish templates from non-template
   configs and do the appropriate thing with each, addressing #11639 in a
   different way.
 * Move resolveTemplate and resolveTemplateForService to configManager.
 * Represent scheduling/unscheduling as a data type (configChanges) for ease of
   testing.
 * Break up getAllConfigs to return a configChanges, with the public
   GetAllConfigs returning the configs themselves.
 * Change the tests to call `processNewConfig` instead of `resolveTemplate`,
   and to adjust for the fact that `processNewConfig` assumes the config is new.

Both `AutoConfig` and `simpleConfigManager` use the `store` type, but they use
_different instances_ of the store and the types of data stored do not overlap:

 * `serviceToConfigs` - `simpleConfigManager` only
 * `templateToConfigs` - `simpleConfigManager` only
 * `loadedConfigs` - `simpleConfigManager` only
 * `entityToService` - `simpleConfigManager` only
 * `adIDToServices` - `simpleConfigManager` only
 * `serviceToTagsHash` - `AutoConfig` only
 * `nameToJMXMetrics` - `AutoConfig` only

This was done to keep the diff understandable.  Once simpleConfigManager is
replaced with its successor, the store can be reduced to just the last two
fields (or just merged into the AutoConfig type).

Note that this contains some "XXX" in
`pkg/autodiscovery/configmgr_test.go`.  These are fixed in a subsequent
commit in this PR.