[workloadmeta] Suppress unset events for unknown entities
This happens sometimes to pause containers, especially in containerd: we
can filter out new pause container events, but once the container is gone we
can no longer get information about its image, so we can't filter out
deleted container events (as we identify them by the image name).

Since this is a bit hard to fix, and to prevent these events from
bothering subscribers, workloadmeta no longer emits events for entites
that were never in the store in the first place.