Don't destroy rtloader on shutdown
rtloader destructor grabs GIL to drop a reference to a python class,
but this may cause a deadlock if a running check holds the GIL too.

Dropping a reference to the check base class is redundant, since we're
shutting down anyway. So we can skip the destructor completely, and
avoid potential deadlocks.