Introduce a new reconcilingConfigManager
This config manager is disabled by default, and enabled by the temporary
`logs_config.cca_in_ad` config flag.  Without that flag, the existing
simpleConfigManager is used.

Currently, the reconcilingConfigManager accomplishes exactly the same
thing as the simpleConfigManager, and the tests demonstrate that.  But
it does so in a way that focuses on reconciling templates and services
on a per-service basis.  This commit should be reviewed as a ground-up
re-implementation of the existing algorithm (simpleConfigManager),
setting the stage for a simple(ish) change in the next commit.