Remove kubeUtil references from logs agent launchers
We're trying to clean up some old code from the kubeUtil, so anything
that fetches pod information needs to be migrated to workloadmeta. This
commit removes all references to it from the docker and kubernetes
launchers, and removes a lot of retrying logic that's now redundant
since pods and containers are guaranteed to be present in the store by
the time the logs agent gets to know about them.