Refactor AD to use a ConfigManager (#11640)
(This PR is broken into smaller commits in the GitHub interface)

* Remove superfluous setTagsHashForService

This call to `setTagsHashForService` occurs when resolving a template
for a specific service.  However, it might overwrite the stored tags
hash for the service as it currently stands, causing AD to miss an
update to the stored tags.  At best, it's unnecessary.

* stop returning tagsHash when resolving a template

* remove tagshash from service.GetTags return

* remove unused tagger.TagWithHash

* Add comments on AD store

* Factor out config management from AutoConfig

This factors out the config management -- handling added/deleted configs
and services -- from AutoConfig into a dedicated type.  The type is an
interface implementation to support its later wholesale replacement
based on a feature flag.

This is a complex diff to read, but at its core it adds new types and
then "lifts" the body of four methods out of AutoConfig and puts them
into the simpleConfigManager type.

 * Change processRemovedConfigs to distinguish templates from non-template
   configs and do the appropriate thing with each, addressing #11639 in a
   different way.
 * Move resolveTemplate and resolveTemplateForService to configManager.
 * Represent scheduling/unscheduling as a data type (configChanges) for ease of
   testing.
 * Break up getAllConfigs to return a configChanges, with the public
   GetAllConfigs returning the configs themselves.
 * Change the tests to call `processNewConfig` instead of `resolveTemplate`,
   and to adjust for the fact that `processNewConfig` assumes the config is new.

Both `AutoConfig` and `simpleConfigManager` use the `store` type, but they use
_different instances_ of the store and the types of data stored do not overlap:

 * `serviceToConfigs` - `simpleConfigManager` only
 * `templateToConfigs` - `simpleConfigManager` only
 * `loadedConfigs` - `simpleConfigManager` only
 * `entityToService` - `simpleConfigManager` only
 * `adIDToServices` - `simpleConfigManager` only
 * `serviceToTagsHash` - `AutoConfig` only
 * `nameToJMXMetrics` - `AutoConfig` only

This was done to keep the diff understandable.  Once simpleConfigManager is
replaced with its successor, the store can be reduced to just the last two
fields (or just merged into the AutoConfig type).

Note that this contains some "XXX" in
`pkg/autodiscovery/configmgr_test.go`.  These are fixed in a subsequent
commit in this PR.

* Gather errorStats related stuff in one file

The errorStats singleton is now referred to both from `autoconfig.go`
and `configmgr.go`, so it makes sense to put its particulars in a
central spot (`stats.go`).

* Move config decryption to a new file

..and soup up the tests a bit, to cover MetricConfig and LogsConfig.

* Only unschedule a config once

..even when the service and template from which it was resolved are both
deleted.

This uncovers the hard truth that the serviceToConfigs and
templateToConfigs potentially grow without bound.  In practice, that is
not a significant issue.  And, this form of config manager will soon be
removed, anyway.

* Fuzz AD ConfigManager, fix uniqueness bug

The fuzzing turned up a bug: if a service is added sharing multiple AD
identifiers with an existing template, the resulting config would be
scheduled multiple times.  The fix is to duplicate the resolvedSet logic
used when adding templates.