hold log files open on windows for closeTimeout
When there is backpressure from the pipeline, we do not want to hold the
file open indefinitely, as this can cause exhaustion of disk space.  But
we also do not want to close the file immediately, as this may cause
data loss if the file has been rotated (see code comments).  The
compromise is to hold the file for `closeTimeout`.  This is the duration
for which the linux tailer holds a rotated file open.