feat(diagnose): move original 'diagnose' command to a subcommand
With the new 'datadog-connectivity' subcommand for 'diagnose' command, we want
to be able to detect typos when launching the command.

For instance if we call 'diagnose blabla', the 'diagnose' command will run,
because 'blabla' is considered as a flag, while we want to display
an error / help message.

By moving the metadata availability checks to a subcommand we are able to
correctly display the help information.

Now instead of using 'datadog-agent diagnoe' we have to use
'datadog-agent diagnose metadata-availability'

Signed-off-by: Nicolas Guerguadj <nicolas.guerguadj@epita.fr>