feat(forwarder_health): improve status message when API Key validation endpoint is unreachable
To validate the API Key, the Agent sends an HTTP Request to a specific endpoint
and checks the response status code.

When the endpoint is unreachable or when the status code is 200 or 403, the API Key
status is set to 'Unable to validate API Key' which is not clear enough that the
error is related to connectivity more than the API Key validity.

Also, this commit introduces more granularity as we can differenciate when we
cannot reach the endpoint from when the status code is unexpected.

Signed-off-by: Nicolas Guerguadj <nicolas.guerguadj@epita.fr>