[AD] Reconciling config manager (#11802)
* add temporary logs_config.cca_in_ad feature flag

* add a test for mapOverLoadedConfigs

* move simpleConfigManager aside

* Introduce a new reconcilingConfigManager

This config manager is disabled by default, and enabled by the temporary
`logs_config.cca_in_ad` config flag.  Without that flag, the existing
simpleConfigManager is used.

Currently, the reconcilingConfigManager accomplishes exactly the same
thing as the simpleConfigManager, and the tests demonstrate that.  But
it does so in a way that focuses on reconciling templates and services
on a per-service basis.  This commit should be reviewed as a ground-up
re-implementation of the existing algorithm (simpleConfigManager),
setting the stage for a simple(ish) change in the next commit.

* Allow AD services to filter templates

With this change, a service can decide which of the available templates
should be resolved and scheduled.

Presently, no Service implementation performs any filtering, so the
behavior does not change.  This functionality will be used in subsequent
commits.