Adding an option to force EC2 instance ID as hostname
In some cases, users might want to force EC2 instance-id as an hostname.
This is needed when autoscaling a VM with a non standard EC2 hostname
(ie: not starting with "ip-", "domu", "ec2amaz-").

Due to legacy logic GCP and Azure already have the priority when
choosing an hostname but not AWS. On EC2 the OS hostname would be
picked. We need to be backward compatible for the hostname detection so
`ec2_prioritize_instance_id_as_hostname` is introduced.